fix: normalize subject teacher foreign keys

+++ EduOS.Core/Entities/Academic/SubjectTeacher.cs
@@ -5,11 +5,11 @@ namespace EduOS.Core.Entities.Academic
 {
     public class SubjectTeacher : BaseTenantEntity
     {
-        public int AcademicYearId { get; set; }
-        public int ClassId { get; set; }
-        public int SectionId { get; set; }
-        public int SubjectId { get; set; }
-        public int TeacherId { get; set; }
+        public long AcademicYearId { get; set; }
+        public long ClassId { get; set; }
+        public long SectionId { get; set; }
+        public long SubjectId { get; set; }
+        public long TeacherId { get; set; }
         public bool IsClassTeacher { get; set; } = false;
 
         public virtual AcademicYear? AcademicYear { get; set; }
